requirements.txt
retry == 0.9.2
requests == 2.25.1
line-bot-sdk == 1.18.0
opencv-python == 4.5.1.48